Major change: app structure, navigation improved * Added constants file * Created file per screen * Added safe area provider for iPhone
