+++ package.json
@@ -20,15 +20,14 @@
     "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.3.tar.gz",
     "react-native-gesture-handler": "~1.7.0",
     "react-native-reanimated": "^1.13.1",
-    "react-native-screens": "~2.10.1",
-    "react-native-web": "~0.13.12"
+    "react-native-safe-area-context": "3.1.4",
+    "react-native-screens": "~2.10.1"
   },
   "devDependencies": {
     "@types/react": "~16.9.35",
     "@types/react-dom": "~16.9.8",
     "@types/react-native": "~0.63.2",
     "babel-plugin-module-resolver": "^4.0.0",
-    "react-native-web": "~0.13.7",
     "typescript": "~3.9.5"
   },
   "private": true
