+++ src/constants/index.ts
@@ -1,0 +1,6 @@
+export default {
+  colors: {
+    darkCyan: '#00848C',
+    tangerineYellow: '#FCCF17',
+  },
+};
