+++ src/navigation/index.tsx
@@ -1,0 +1,24 @@
+import constants from '@/constants';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+import React from 'react';
+import BottomTabsNavigator from './BottomTabsNavigator';
+
+const Stack = createStackNavigator();
+
+export default function Navigation() {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator
+        screenOptions={{
+          headerTintColor: constants.colors.darkCyan,
+          headerTitleStyle: {
+            fontWeight: 'bold',
+          },
+        }}
+      >
+        <Stack.Screen name="Root" component={BottomTabsNavigator} options={{ title: 'Alvo' }} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
